final update : ready to build

+++ lib/screens/admin/list_exicutives/functions.dart
@@ -1,7 +1,8 @@
 import 'dart:developer';
-
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:connectivity/connectivity.dart';
 import 'package:detail_dex/screens/admin/bloc/admin_bloc.dart';
+import 'package:detail_dex/widgets/no_network_widget/no_network_widget.dart';
 import 'package:detail_dex/widgets/snackbar_widget/snackbar.dart';
 import 'package:flutter/material.dart';
 
@@ -71,13 +72,54 @@ alertPassshow(
 }
 
 deleteExicutivedetails(String id, context, bloc) async {
-  Navigator.pop(context);
-  bloc.add(AdminDeleteLoading());
-  final CollectionReference firedata =
-      FirebaseFirestore.instance.collection('Exicutives');
+  final connectivityResult = await Connectivity().checkConnectivity();
+  if (connectivityResult == ConnectivityResult.none) {
+    Navigator.of(context).push(MaterialPageRoute(
+        builder: (ctx) => const NoNetworkWidget(
+              isSplash: false,
+            )));
+  } else {
+    Navigator.pop(context);
+    bloc.add(AdminDeleteLoading());
+    final CollectionReference firedata =
+        FirebaseFirestore.instance.collection('Exicutives');
 
-  await firedata.doc(id).delete();
-  Navigator.pop(context);
+    await firedata.doc(id).delete();
+    Navigator.pop(context);
 
-  snackbarSucess(context, 'Deleted successfully');
+    snackbarSucess(context, 'Deleted successfully');
+  }
+}
+
+Stream<int> countTodayExicutiveClients(String exicutive) async* {
+  DateTime today = DateTime.now();
+  DateTime todayDate = DateTime(today.year, today.month, today.day);
+  try {
+    final messagesSnapshot = FirebaseFirestore.instance
+        .collection('Details')
+        .orderBy('datetime', descending: true)
+        .snapshots(includeMetadataChanges: true);
+
+    await for (final messages in messagesSnapshot) {
+      List todaygot = [];
+      final detailslist = messages.docs.map((e) {
+        Map<String, dynamic> data = e.data();
+
+        return data;
+      }).toList();
+      for (var i in detailslist) {
+        Timestamp timestamp = i['datetime'];
+        DateTime datadate = timestamp.toDate();
+        DateTime databasedate =
+            DateTime(datadate.year, datadate.month, datadate.day);
+        if (i['exicutiveid'] == exicutive && databasedate == todayDate) {
+          todaygot.add(i);
+        }
+      }
+      yield todaygot.length;
+    }
+  } catch (e) {
+    log('Error fetching details: $e');
+    yield 0;
+  }
 }
